requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
cfgv==3.1.0
click==7.1.2
distlib==0.3.0
Django==3.0.6
django-copyright==1.0.0
django-cors-headers==3.3.0
djangorestframework==3.11.0
filelock==3.0.12
flake8==3.8.2
identify==1.4.17
importlib-metadata==1.6.0
Markdown==3.2.2
mccabe==0.6.1
nodeenv==1.3.5
pathspec==0.8.0
pipenv==2020.5.28
pre-commit==2.4.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
virtualenv==20.0.21
virtualenv-clone==0.5.4
zipp==3.1.0
